+++ src/as_display/mod.rs
@@ -1,0 +1,7 @@
+mod as_debug_display;
+mod debug_display;
+mod debug_displayed;
+
+pub use as_debug_display::AsDebugDisplay;
+pub use debug_display::DebugDisplay;
+pub use debug_displayed::DebugDisplayed;