+++ src/result/mod.rs
@@ -1,0 +1,5 @@
+mod opaque_dbg;
+mod type_debug;
+
+pub use opaque_dbg::OpaqueResultDbg;
+pub use type_debug::ResultTypeDebug;